requirements.txt

#list the package names to be installed
python-dotenv
pandas